OpenKF for our case is building and predicting
Our prediction step matches the results of the OpenKF example and is
passing that portion of the test. It is failing the update step, but
this might not be a problem as we are not adding the measurement noise
to the augmented state vector, as the OpenKF example does. Instead we
are treating the measurement noise analytically. This might cause the
small differences in the results, but I need to run through the math to
confirm that.