Add first version of unpacker for Si data (#251)
* Fix error reporting in AtHDFUnpacker

The open function now properly returns the
expected number of events, so the value returned
does not need to be divided by 2 anymore.

* Refactor methods that just open and count entries in a group

* Refactor FRIB unpacker to support multiple digitizers

* Allow digitizers to have different freq

* Unpacking all pads for Si unpacker

Still need to actually unpack the Si detectors.
At the moment appears to be reading in the FRIB data correctly and the
pads, but they have no data in them in my test file so hard to confirm.

* Poll object type before reading timestamps.

* Add Si to aux channels

* Run clang format

* Add test viewer for frib_physics

At the moment appears to be reading all the data correctly, but nothing but FRIB digitizers have data so hard to confirm

---------

Co-authored-by: anthoak13 <anthonya@frib.msu.edu>